test: 자동차 이름 검증 테스트 이전
